Fixes package uninstall deletion logic
Bug: 311250325
The DeleteUninstallJobService job was deleting more data than it was
supposed to. Additionally, the broadcast receiver was not handling the
deletion of measurement data correctly. It was not resetting some of the
fields that it should have been. To fix these issues, the code has been
refactored to use the deletion registration API logic. This logic takes
care of all of the edge cases. This also fixes the issue with the
missing reference ids for noise reports. These ids are used for data
deletion on cascades.

Test: atest AdServicesServiceCoreUnitTests
Test: atest AppDeletionIntegrationTest
Change-Id: Ic9ad751f9558c7bafc21e9b336698b6b663cae14