Change Flag Name Constants to Local Fields
The update-api tooling pulls a reference to FlagsConstants if the key is directly used in the annotation. However, it correctly pulls in the key value from class-local constants. So reading the value of the key into a local constant which is then used within the annotation.

Bug: 314002642
Test: m
Change-Id: I94e92f5bede8b5394642c0a6bbaaad1b8c327cc2
Merged-In: I94e92f5bede8b5394642c0a6bbaaad1b8c327cc2