Define Mode on ProcessLifeGuardRule.
Initially, that rule was always failing the current test when it
caught an exception on background threads, but that behavior made
it hard to fix test failures (as often the exception was thrown by
a previous test).

Now it defines 3 modes (FAIl, IGNORE, and FORWARD), and
AdServicesUnitTestCase uses IGNORE by default.

Test: adb shell setprop debug.adservices.test.postTestDelay 500 &&\
  adb shell setprop debug.adservices.test.postTestThrownFrequency 2 &&\
  atest AdServicesServiceCoreUnitTests:AppManifestConfigMetricsLoggerTest
Test:  m -j AdServicesTestUtilityTests && adb install -t -g $OUT/testcases/AdServicesTestUtilityTests/*/AdServicesTestUtilityTests.apk && adb shell am instrument -w --no-hidden-api-checks -e class com.android.adservices.common.ProcessLifeguardTestSuite -w com.android.adservices.testutil.tests/androidx.test.runner.AndroidJUnitRunner

Bug: 303112789
Bug: 285014040
Fixes: 313554699

Change-Id: Ic9abdb1bb43f6306f4468a3e3bc7d73b57af6e51