Fix Tests not Running on Mainline Presubmit [Infra]
Fix or Clean up the wrong configuration for mainline Presubmit. See below and the bug for the details and docs.

For a test with AndroidTest.xml, the tag should be set in the .xml file instead of .bp. Such tests with incorrect configuration will not show up in the mainline presubmit test mapping.

For example, in a U-mapping , below test cases are missing:
  - CtsAdIdEndToEndTest
  - CtsAppSetIdEndToEndTest
  - CtsAdServicesCobaltTest

Bug: 275573543
Test: Will check the same U/S mainline mapping to see if tests are added.
Change-Id: Ie5d12305431ab51e12b66ff37875f89292414aad