Clean Up Redundant Mainline Presubmit Tag [Gateway]
For a test with AndroidTest.xml, the tag should be set in the .xml file instead of .bp. Such tests with incorrect configuration will not show up in the mainline presubmit test mapping.

All Gateway related tests have the correct configuration and run on the Mainline presubmit. Per request from Mainline Team (See the bug), this CL is to clean up the redundant tags in the .bp files.

Bug: 275573543
Test: Will check the same U/S mainline mapping to see if tests are still included.
Change-Id: Iccd85fc7892e5c38eb8809f549c58875bdfc5e8d