Fixes attribution running more times than needed
Bug: 313439580
Attribution job was running more times than it was supposed to. The
reason was that every time it finished processing, the job was
rescheduled to be executed immediately instead of waiting for the
content provider notification. However, it was not causing any data
harm, but the phone would be using more resources than needed.

Test: atest AttributionJobServiceTest
Change-Id: Ia07d7ef9b5946f8f623c0fb4dd22ce30bf19be26