cli: track if CA is debuggable in main table
This builds upon ag/25466937 to track the debuggability of a custom
audience in the main CA table, as well as the background fetch table.
Duplication matches the existing pattern with data required for both
background fetch and main CA. This bit will be read by CLI commands
for listing / viewing custom audiences (http://ag/25631963).

Also cleanup all instances of "/*debuggable=*/" in the code.

Test: atest AdServicesServiceCoreUnitTests
Bug: 308391017
Change-Id: I34df7b6f289b3527272902ad9cc8e7ce3b1bbfd7