Fix uninstalling ctssdkprovider after the test runs
SuiteApkInstaller is not able to uninstall the Sdk-Provider.apk because the test
app that requires it is still installed at that point. This means the apk would
stay on the test device after the test ends.

This preparer, during tear down, runs after SuiteApkInstaller and will finally
uninstall the provider, leaving the device in a clean state.

Bug: 314294893

Test: atest  SdkSandboxManagerTests SdkSandboxManagerDisabledTests CtsStrictJavaPackagesTestCases

Change-Id: Ie53b9002670263349f776faa25b2a019fa1123ac
Repro: run the sandbox tests
Repro: adb shell pm list packages --match-libraries -f  | grep ctssdkprov
Repro: it should print nothing after the test is done