Pass the root cause exception to the thrown exception on AppSearch failure
We can observe a small number of crashes happening when attempting to write data to AppSearch. However, the code catches the underlying exception, logs it to logcat, and then throws a generic RuntimeException. As a result, the information on the underlying root cause is lost when viewing crash logs, making it difficult to fix. This CL ensures that all AppSearch writes pass the underlying exception that's thrown to the generic RuntimeException.

It also increases the timeout that we have for AppSearch, since it seems there's a small number of requests that go past that timeout. In order to be able to control this timeout without requiring a module rollout, we can read the timeout values from PhFlags.

It also adds some cosmetic improvements, such as switching away from the deprecation version of LogUtil.e() that was being used.

Bug: 316920565
Test: atest AdExtServicesServiceCoreAppSearchUnitTests
(cherry picked from https://googleplex-android-review.googlesource.com/q/commit:5d635bfd7b83c00deca1ee8caaa85368fe495d67)
Merged-In: I4ad0cf0a2edea32a476fcd0bd20c97f28761f3bc
Change-Id: I4ad0cf0a2edea32a476fcd0bd20c97f28761f3bc