Move StatsD logging to a separate class
A new singleton instance class, SdkSandboxStatsdLogger, is created to
perform all logging to StatsD in it. This is done to declutter
SdkSandboxManagerService which is currently used for this matter. All calls
are now redirected from the service to the logger -- logger is only
accessible in System Server.
Additionally, all logging unit tests are updated: calls to StatsD are
now verified in the logger test class, and
SdkSandboxStatsdLoggerUnitTest verifies that the logger is called and
logging object parameters (e.g. of SandboxLatencyInfo) are set
correctly.

Bug: 314751800
Test: atest SdkSandboxManagerServiceUnitTests
Change-Id: I912ba6f76db1ffc414e177394f243bcd43c78990