Increase AppSearch read timeout since it appears that p99 > 500ms
See https://screenshot.googleplex.com/6GsEJoXt4gSAUA2, it shows that the call to AppSearch GlobalSearch is currently taking just over 500ms. This will result in a timeout and no data being returned. Bumping up the timeout to lessen the chances of that happening.

Also found some other instances where we were logging the original thrown exception to logcat and re-throwing a generic RuntimeException without passing in the original as the cause, fixing those as well.

Bug: 316920565
Test: atest
Change-Id: I6f912308c779cedeb4969dd4e2c7ca3187d9beb7