cli: Add view and list custom audiences commands
Call DAO methods to list and view custom audiences for a given owner
and buyer. Only CAs created in a debuggable context are visible.

Also adds an initial version of an argument parser, where syntax is
as follows: --key=value. Arguments can be marked as required, with
additional exception handling added for duplicate or null/empty args.

Test: atest com.android.adservices.service.shell
Bug: 318496217
Change-Id: I42929a05f84949aab2d199b8b3f4f994f478439d