Propagate exception within BlockedTopicsManager for visibility in crash reports
Similar to bug 316920565, the BlockedTopicsManager code catches the exception thrown from within AppSearch (and other functions), logs it to logcat, and replaces it with a new RuntimeException containing a generic error message. This results in the root cause not being shown in the crash logs, making it impossible to debug further. This CL ensures that the root exception is passed to the generic RuntimeException as a cause.

Also, it appears that some of the tests in BlockedTopicsManagerTest were passing by accident, because they were insufficiently asserting. The test was expecting a RuntimeException to be thrown by the code, but it didn't assert anything other than the exception type. When the test executes, the code actually fails because of a different error with a SecurityException, but the assert passes because SecurityException is a subclass of RuntimeException. Fixing it so the test looks for additional appropriate attributes of the thrown exception.

Bug: 321720113
Test: atest com.android.adservices.service.topics.BlockedTopicsManagerTest

Change-Id: I068cdeb919bdcafbdc7167c57b4bf381360c3657