Fixes flakyness on test class DebugReportingJobServiceTest
Bug: 322103786
Fixes flakiness on DebugReportingJobServiceTest#onStartJob_shouldDisableJobTrue. The issue is that it is also doing an additional validation to determine if a static method is not called. As tests can run concurrently, this can't be a guarantee. The test is already checking if the job gets canceled which is more than enough as that is the end of the flow.

Test: atest DebugReportingJobService --iterations
Change-Id: Ia4dbe0ab2a102e35207e144737ff7596e9ca5391