Fix failing AppSearchConsentMigrationTest on AOSP CF target
The AppSearchConsentMigration test has been failing on git_main, on the aosp_cf_x86_64 target 100% of the time. The root cause was that the data from the test helper app was being written to AppSearch was being configured to be readable from the AdServices modules on GMS devices, but not AOSP. As a result, when AdServices queried AppSearch for notification data, it returned nothing, resulting in the test failing. Adding the SHA of the AOSP AdServices apk to the helper app fixes the problem.

Bug: 308240435
Test: atest CtsAdServicesHostTests:AppSearchDataMigrationHostTest#testAppSearchConsentMigration
Change-Id: I7e30d7a9deb8a33b67c50f16d7a90975bc1086ec