Ignore 5 buyer CB test instead of marking as flaky
B&a doesnt support 5 buyers right now. So this test is bound
to fail more often than not. Marking it as just Flaky still
runs the test and pollutes our dashboard from finding useful
failures.

The other option is to delete the test altogether but once B&a
supports 5 buyers by Q2, we want to re-enable this test

Test: adb shell am instrument
Bug: 322323696

Change-Id: I4831f0a509927dd6dbcfa9f7aa19c97a83a30239