Warmup B&a e2e tests
We have flakiness in our e2e tests because sometimes B&a throws error
when connected to after a while.
95 buyer_frontend_async_client.cc:49] SendRPC completion status not ok:
UNAVAILABLE: recvmsg:Connection reset by peer

A test execution immediately after the failure often passes.

Hence, warming up with a pre-call and waiting for a couple of seconds
might be beneficial to reduce flakiness

Test: adb shell am instrument

Change-Id: I0056e91830f6479a413c03189a4480c44996587c