[DON'T BLOCK] Test ownership migration rules
This CL is created as a best effort to migrate test targets
to the new android ownership model. If you find incorrect or unnecessary
attribution in this CL, please create a separate CL to fix that.

For more details please refer to the link below,
go/new-android-ownership-model

Bug: 304529413
Test: N/A
Change-Id: Ic4cd8d76e032b51c084207d7da5836d273d701bc
Merged-In: Ic4cd8d76e032b51c084207d7da5836d273d701bc