API to get the effective target SDK Version
Added a new API to fetch the effective target Sdk Version for an app.
This API will be used by platform in future to fetch ETSV and assign the corresponding selinux restriction to sdk sandbox process.

This is not a clean cherry-pick, the API in this branch is hidden and the hide annotation will be removed once the flag in the main branch is ramped up to 100%

Bug: 271547387
Test: atest SdkSandboxManagerServiceUnitTests
Test: atest SdkSandboxRestrictionManagerUnitTest

Merged-In: I3be18cb57d984381db6dbd4e0f4864a5992e5301
Change-Id: I3be18cb57d984381db6dbd4e0f4864a5992e5301
(cherry picked from commit 476a272d7effef5bb5a9d0490e57c993e09344ff)