Add latency logging for SdkSandboxController activity APIs
Latency logging is added for registering and unregistering of sandbox
activity handlers in SdkSandboxController. Latency data is collected in
controller, and then routed to SdkSandboxManagerService via
SdkSandboxLocalSingleton.
Additional check is added to SdkSandboxStatsdLogger to convert sandbox
uids to client app uids for cases when API is called from the sandbox
process.

Design doc: go/rbs-activity-metrics
Bug: 300102108
Test: atest SdkSandboxFrameworkUnitTests

Change-Id: I91861ecdca69f632c76e7cb9b9b06725714d63ba