Add latency logging for SdkSandboxActivityRegistry APIs
Latency logging is added for SdkSandboxActivityRegistry APIs related to
Sandbox Activity creation, which are registering, unregistering activity
handlers and notifying about activity creation. Latency data is
collected in the registry, and then sent to SdkSandboxManagerService by
using SdkSandboxLocalSingleton. Injector is added to the registry for an
ease of logging and testing.

Design doc: go/rbs-activity-metrics
Bug: 300102108
Test: atest SdkSandboxFrameworkUnitTests SdkSandboxManagerServiceUnitTests SdkSandboxUnitTests; statsd_testdrive 735

Change-Id: I10f02bf2a078d206109f9ca3b248ef9decd6312e