Add total sandbox activity creation latency logging
Total sandbox activity creation latency is logged by putting an
additional extra in the intent in SdkSandboxManager. This extra
represents the time sandbox activity creation was initiated. It is then
consumed in SdkSandboxActivityRegistry, and, if the activity creation
was successful, total call latency is reported to System Server to be
sent to StatsD. This additional extra is removed from the intent after
the first notification call to avoid logging the wrong time in case of
activity recreation.

Design doc: go/rbs-activity-metrics
Bug: 300102108
Test: atest SdkSandboxFrameworkUnitTests SdkSandboxManagerServiceUnitTests SdkSandboxUnitTests; statsd_testdrive 735

Change-Id: I7cc357c9b6a8aaac57602bd853bced11b3daef09