Fix skipping for SDK sandbox disabled on tests that use @BeforeClass
All classes except for SdkSandboxWebViewClientTest run using @ClassRule
rather than just @Rule. This means loadTestSdk is run in @BeforeClass.

Using an assumption in @BeforeClass leads to this 'failure' since all
the tests are skipped and instrumentation fails reporting that not
enough tests were run.

Therefore, we avoid using assumeTrue in loadTestSdk, and instead check
if the test SDK executor is null in assertSdkTestRunPasses and assume
then.

Test: atest CtsSdkSandboxWebkitTestCases (all 247 tests run and skipped)
Bug: 317676206
Change-Id: I56eed325e43a2b0ceefcc3cd91ca7a1a4f5d90fc