Ensure sandbox kill on app uninstall
Currently, the sandbox does not die when the app is uninstalled because
in sandbox kill flow, app process name is attempted to be fetched from
PackageManager. However, since the app no longer exists, getting the
process name fails and sandbox is not killed.

This change stores the sandbox process name when it is first created.
Later, in sandbox kill flow, the stored sandbox process name will be
used to stop the sandbox service.

Bug: 329265306
Test: atest SdkSandboxLifecycleHostTest#testSdkSandboxIsKilledOnAppUninstall
Change-Id: I0ed9efddd5e49c30f575d8c98a64bb9602554916