cts: Ask OS for free port for MockWebServer
Force the kernel to give us a free port instead of a hardcoded :83838
which may still be active from previous test runs. This should improve
stability of MockWebServer tests. Measurement's tests which require are
still hardcoded.

This CL introduces a feature to have variable substitutions in mock
paths (see scenario-test-010.json for an example) as these can no longer
be hardcoded.

Bug: 315327589
Bug: 333408258
Bug: 329550649
Test: atest CtsAdServicesDebuggableDeviceTestCases \
	&& atest CtsAdServicesRootTestCases
Change-Id: I9b68968dafb972ead38a9e8e8be44de800b4c8be