Turn on customized sdk context by default
Customized SDK context is needed for per-sdk storage and full-screen
activity feature of sandbox. As such, we won't enable sandbox without
enabling this flag.

Setting the default value to true for now to unblock test failures. In
follow up CL, will clean up the flag from code since it's expected to be
turned on U+ device all the time.

Bug: 329867471
Bug: 323811054
Test: presubmit
Change-Id: Ibfe9d9703cf276cc7dca4ef46b555d5981bc1086