Lifecycle test: avoid testing app is killed when SDK updated
We don't necessarily need the app to be killed when an SDK update
occurs, provided that the SDK sandbox is killed. Testing that the app
process is also not there is not necessary, and flaky since AM may
restart the app process.

The behaviour which kills an app on SDK update belongs to package
manager, and is tested separately under StaticSharedLibsTests
testSamegradeStaticSharedLib_killDependentApp.

Bug: 332003149
Test: atest SdkSandboxLifecycleHostTest#testSdkSandboxIsKilledOnLoadedSdkUpdate
Test: atest SdkSandboxLifecycleHostTest#testSdkSandboxIsKilledForNonLoadedSdkUpdate
Change-Id: Id846898a10639d22ddb3a7d8b76dab51a5347001