Fixes flakyness on AsyncRegistrationQueueRunnerTest
Bug: 333879224
The method
testRecordsExceedMaxRegistrationsPerJob_returnSuccessWithPendingRecords
failed with timeout while trying to perform a network call. The test is
main goal is to test the behavior when it process all records, the
network call just happens to be on the same flow. The network call has
been mocked to avoid performing a network call that could result in a
timeout.

Test: atest AdExtServicesServiceCoreMeasurementUnitTests:com.android.adservices.service.measurement.registration.AsyncRegistrationQueueRunnerTest --iterations
Change-Id: Idb4f99397a2bb9ae69450d391c0e82daa769a28a