Introduce AdServices logging usage rule.
Introducing some components that will be used as part of a new
AdServicesLoggingUsageRule which will be responsible for scanning
supported log calls and ensuring all calls have been verified through
annotations. See bug for detailed rationale.

The goal is to have the rule run as part of all classes that extend
AdServicesExtendedMockitoTestCase. However, until that is possible, it
will be a standalone rule that can be used directly in test classes
where needed. The rule has been designed to easily configure any log
calls. Initially, the focus is on ErrorLogUtil calls.

Note: This CL is only meant to showcase the logic that resides in
AdServicesLoggingUsageRule. Much of the logic to actually verify the
calls has been abstracted out and will be fleshed out in subsequent
CLs.

Bug: 337041767
Test: atest AdServicesSharedLibrariesUnitTests
Flag: TEST_ONLY
Change-Id: I90fc603a575e532dfb4fd79e9f3db75f267be038