Introduce generic log verification framework for supporting ErrorLogUtil calls.
We should be able to easily configure new log types in AdServicesLoggingUsageRule. This CL introduces a generic design to accomplish that and uses it to support verification of ErrorLogUtil calls. Goal of the CL is to support explicit verification of both ErrorLogUtils.e(Throwable, int, int) and ErrrorLogUtils.e(int, int) calls in unit tests. Multiple annotation support and other features to simplify log verification in tests will be implemented in other CLs.

Bug: 337042949
Test: atest AdServicesSharedLibrariesUnitTests AdServicesTestUtilityTests
Flag: TEST_ONLY
Change-Id: I66b35b71b41e0412faaaf351b1f32607c1dd9172