Remove Job ID in JobSpec and Enforce it in JobPolicy
Since JobPolicy is a proto and proto recommends to set all fields as optional, it cannot enforce the JobPolicy to have a job ID. Instead, enforce it in the Builder of JobSpec. As a result, job ID is no longer needed in JobSpec as it can be achieved in the JobPolicy, because JobPolicy is a required field and required to have the job ID.

Also modified MDD to reflect this change. And this change is required for M06.

Bug: 323029946
Test: atest JobSpecTest PolicyJobSchedulerTest
(cherry picked from https://googleplex-android-review.googlesource.com/q/commit:ff658cc53d4772445dfb7972a7a07a5659f22fe0)
Merged-In: I29c0db79ec1ede4cb2583f731b62cd4940cddd84
Change-Id: I29c0db79ec1ede4cb2583f731b62cd4940cddd84