Deprecate PhFlag usages in SPE Jobs
The constraint update in SPE is via the proto-based Policy for all jobs and will be evaluated when the SPE JobScheduler schedules the job. In the Job Instance, it should just use the default value and is redundant and risky to depend on other PhFlags.

Also update the return value for onStopJob() in BackgroundFetchJob to true, to be consistent
with BackgroundFetchJobService.

Bug: 335489810
Bug: 335489462
Test: atest BackgroundFetchJobTest EpochJobTest
Change-Id: I4ee3039feb63905290449314eea8a18e1d348731