Fix CustomAudienceTest cleanup and add verbose logs
This test fixes an error in CustomAudienceTest where only
testCustomAudience_doesNotPersistAfterAppUninstallAndReinstall(), but not
testCustomAudience_doesNotPersistAfterAppUninstallAndReinstall(), awaited
app uninstall in test cleanup before proceeding to the next test.

This produced a race condition where, if the latter test ran first, the
former test could try to install an app that was still installed,
causing the assertion verifying the install in installApp() to fail.

This CL adds the same check to the latter test, assuring the race
condition will not emerge regardless of the order tests are run.

Additionally, it adds verbose logging throughout the tests to improve
debugging of this and other such errors in the future.

Bug: 340383282
Test: atest CtsAdServicesCustomAudienceTests
Flag: TEST_ONLY
Change-Id: Iebf4983fa37faf2e34b2d00486ae9ee4d55b7ad5