Dettached TestSyncCallback from SyncCallback - step 1
The current SyncCallback was created with the intention to be used
on "production" code, but in reality it's unlikely to (as production
code should be async) and added more confusion in the class /
interfaces names.

So, this CL dettaches TestSyncCallback from SyncCallback and rename
the latter to ProdSyncCallback; the next CL will remove
ProdSyncCallback and rename TestSyncCallback to SyncCallback
(it's better to do in a separate CL to avoid git diff confusion),
while a 3rd CL will clean up the code.

Bug: 337014024
Test: atest AdServicesSharedLibrariesUnitTests
Flag: TEST_ONLY
Change-Id: I7c557ae656548b2da09415d978c2a797a0250483