Add Error Message for ServiceUnavailableException
This change is like a revert of of commit 1c7808141b15ef523b864d6fa418c549cf92fa2f. The reason is ServiceUnavailableException is not exposed as an API class so SDK cannot get the except type. Instead, the caller receives an ISE without original error message.

Flag: EXEMPT (bug 337358613)
Bug: 343478008
Test: m
Change-Id: I949de5dd5922b45873c0ee02b4dfed9e95cb4861