fixing flake in mockwebserver rule
Earlier the Mock Web Server rule was reserving the port by creating a socket, getting its port and closing this socket and then the web server started with the same port.
Now, we are just starting the mock web server with port 0. It internally reserves a random port.

Bug: 332959958
Test: atest CtsAdServicesDebuggableDeviceTestCases CtsAdServicesRootTestCases
Flag: TEST_ONLY

Change-Id: Ic2312ef009d1da111c81b867876a5b5c21efbb58