Parameter check for service allowlist
In the current implementation to check if a service can be started or
bound in the sandbox process, we check for intent#package and
intent#component#package. We should instead check if any one of the
package is being set.

This CL makes changes to accomodate the above requirement, and adds some
checks which can fails starting or binding to a service if
1. intent#action is null
2. intent#package is null or intent#component is null

Bug: 341652218
Test: atest SdkSandboxManagerServiceUnitTests
Flag: com.android.sdksandbox.flags.service_restriction_package_name_logic_updated
Change-Id: I65d1c82449eb251ba3cd2a4904e5e4ec69eecc6b