Final (?!?!?!?!?) improvements on SyncCallback - step 6
Removed setCalled() from SyncCallback.

SyncCallback comes in 2 flavors: those that return a result and those
that don't.
Currently, SyncCallback defines a setCalled() method, so the latter
must override it to thrown UnsupportedOperationException, which is
kind of hacky.

This change split setCalled() into a ResultlessSyncCallback
interface, which solves that problem.

Test: atest AdServicesSharedLibrariesUnitTests --iterations 10 --test-filter '.*Sync.*'
Test: atest AdServicesSharedLibrariesUnitTests # Case, the name is Justin Case.

Flag: TEST_ONLY
Bug: 337014024

Change-Id: Ia7582034ae70261ecc5f2b8100a6105c8a9f1108