Fix test that was broken by an earlier refactor to set logcat tags
In commit d58f5cb738699d5a545abadfdc41ac0083af229a, a change was made to remove the explicit setting of the logcat tags in the test method and move it to an annotation instead. The annotation processor runs before the setup method, processes both the class level and method level annotations, and sets the logcat tags. However, the setup method includes a device reboot, which causes the logcat tag verbosity to return to the default value. As a result, the logs are missing some lines that the test is dependent on.

This is one case where the annotations cannot be used, because of the intermediate reboot.

Bug: 344852280
Test: atest CtsAdServicesHostTests:AppSearchDataMigrationHostTest
Flag: TEST_ONLY

Change-Id: I7ace0389b9c1fdc916f44b833af283d63b6d9a78