Revert "Temporarily Remove RequiresAndroidServiceAvailable Annotation"
This reverts commit 65de4d3167514f09f736a45bb85b9017fc1a4d7a.

Reason for revert: Previously I saw some invocations skipping the test due to the annotation. However, they could be actual scenarios, such as aosp targets, that need to skip the test. I put two invocations of AdIdCompatibleManagerTest (which keeps the annotation on the class level) in the bug. They are one presubmit invocation and one postsubmit invocation. Both constantly passes the test and show 1 AdIdProvider in the logcat.

Bug: 328133731
Change-Id: Ie4dc6611ec26ea793a6874636d3af373529c85ed