Create separate annotations for ErrorLogUtil verification.
We will soon be creating a new class level annotation to provide default
args for @ExpectErrorLogUtilCall annotations. The problem is current
annotation handles both variants (one with exception and one without).
Therefore, it makes it difficult to discern when it's appropriate for
the annotation to inherit the throwable parameter from the class level
default param annotation.

For a cleaner solution of default parameters (upcoming in the next CL),
this is a refactor to create a new @ExpectErrorLogUtilWithExceptionCall
to handle ErrorLogUtil calls with exception. As a result of this, a new
log verifier was also created to process this annotation; however
`ErrorLogUtilCall` is still modeled to handle both variants as it's an
internal implementation detail not realized by test authors.

Bug: 337963152
Test: atest AdServicesTestUtilityTests AdServicesSharedLibrariesUnitTests AdExtBootCompletedReceiverTest AdServicesBackCompatInitTest PackageManaterCompatUtilsTest
Flag: TEST_ONLY
Change-Id: I22f4da2aa1c9346ff42b41a72f3a0ca9ed107d7d