Replace mostlyEqual check with atLeast check
On small percentage of runs ( < 1%), data written on device exceeds our
error of margin. Instead of trying to guess the margin percentage, we
are opting to just check that we have written at least the amount of
data in the test.

Very unlikely for device to suddenly write 10 MB of data randomly. So a
jump of 10 MB signifies attribution happening due to our test.

Bug: 333356494
Test: atest SdkSandboxStorageHostTest
Change-Id: Ib9a5d9cdc4f5a7c328cbfa14e326ee76f7a09332