Wait for secondary user to be removed completely
If it takes too long to remove a user, it could delay loadSdk call in
following test run, resulting in time out for it.

Considering all tests that create new user are large test and run on
post-submit only, it should be fine to wait for user removal before
starting the next test.

Bug: 334937390
Test: atest SdkSandboxStorageHostTest
Flag: TEST_ONLY
Change-Id: I174045b442715be74dd63a6361d6a85691891c47