Fix getCause() Null issue for AdId and AppsetId CTS GO tests
AdId and AppsetId have updated to throw and assign the error to callback directly and Topics API still uses the old implementation to throw the exception on the API call. Therefore, Topics API throws an exception with the cause while AdId and AppsetId CTS tests throw the exception with no cause.

Flag: TEST_ONLY
Bug: 347266702
Bug: 347145884
Test: Tested with GO device with "atest AdIdCompatibleManagerTest AppSetIdManagerTest"
Change-Id: I191a3cd9c1e713cb8cb906f476970418f0b27bdf