Removing old checks for ReportAndRegisterEventImplTest
There was a mistake in commit ebc7ae85493284f694445dccb35c9ae6e0b49010 that the old type of checks were not removed.

And in the test log it did show the seller callback has happened and was caught by the AnswerSyncCallback, but the test failed earlier on the old checks.

Flag: TEST_ONLY
Bug: 346551750
Test: atest ReportAndRegisterEventImplTest
Change-Id: Ic24d779533f8a3442cb6968453de0ec4a5ae2e66