Ignore tests failing because of room upgrade
Due to a recent upgrade in room, some of our test where we were inserting null in the database and expecting an exception to be thrown are now failing.
We should follow up with the room team. Ignoring
this test in this CL and we will follow up with the room team in a
seperate bug/CL.

Test: CustomAudienceDaoTest
Flag: TEST_ONLY
Bug: 347286338
Change-Id: I0c56e5d2351f5c2f0cf497f1e2856731724759df