Switch to user 0 before running storage host tests
SdkSandboxStorageHostTest runs with the assumption that the current user
is 0. This can fail if the tests starts in a different user (for example
if a previous set of tests left the device in this state). Switch to the
primary user before running tests to avoid this issue.

Bug: 347813282
Flag: TEST_ONLY
Test: adb shell pm create-user test && adb shell am switch-user 10 &&
atest SdkSandboxStorageHostTest. The tests fail unless this change is
applied.

Change-Id: I7cad980afc649bc531c236cd540ef0674b00ef66