Replace ps -A with getRunningAppProcesses in SdkSandboxUnixStreamTest
The ActivityManager.getRunningAppProcesses API does not rely on
executing shell commands. It is easier to debug and hopefully it
will reduce the flakiness of the test.

For pre-V devices the getRunningAppProcesses API does not return the
correct list because test instrumentation for SDK sandbox processes is
not set up correctly. For that case, we rely on the adb shell command
pgrep instead of the ActivityManager API.

Also, add a condition to make sure the sandbox instrumentation tests
only run in SDK sandbox supported environments.

Flag: TEST_ONLY
Bug: 344162120
Bug: 341781690
Bug: 344498720
Test: atest SdkSandboxUnixStreamTest
Change-Id: Ia2280be9813709be33b3315f121e325cfaab2365