Parameter check for service allowlist
In the current implementation to check if a service can be started or
bound in the sandbox process, we check for intent#package and
intent#component#package. We should instead check if any one of the
package is being set.

This CL makes changes to accomodate the above requirement, and adds some
checks which can fails starting or binding to a service if
1. intent#action is null
2. intent#package is null or intent#component is null

Bug: 341652218
Test: atest SdkSandboxManagerServiceUnitTests
Test: Manual test to ensure that the flag was available in MTS
Flag: com.android.sdksandbox.flags.service_restriction_package_name_logic_updated

This reverts commit 497edf8160e8a6fb606fbe0e505407a85bbb469f.

Change-Id: Icbcf9d57f239d653fa7c63ab197ce02c8dd193a8