Ignore tests failing because of room upgrade
Due to a recent upgrade in room, some of our test where we were inserting null in the database and expecting an exception to be thrown are now failing.
We should follow up with the room team. Ignoring
this test in this CL and we will follow up with the room team in a
seperate bug/CL.

Change-Id: I59c8b6ad883ff55411f7f6d598cab72ce8709f95
Merged-In: I0c56e5d2351f5c2f0cf497f1e2856731724759df
Test: CustomAudienceDaoTest
Flag: TEST_ONLY
Bug: 347286338