Split secondary user inprocess test into separate test module
Note that the new module is not marked as CTS, because CTS tests already
support running tests as a secondary user.

The new test module is only added to the postsubmit tests and not the
presubmit ones because of test mapping SLO requirements
(go/test-mapping-slo-guide). It can be graduated to presubmit after
sufficient soak time in postsubmit.

Flag: TEST_ONLY
Bug: 349480881
Test: atest --test-mapping packages/modules/AdServices/sdksandbox:presubmit
Change-Id: I254c7bf8fd1dee0ad56cec090d912ddf528060d3