Add jarjar rules for tests
The aconfig flags are delivered as static lib as part of `service-sandbox` and `framework-sandbox` module and the flags are delivered in post jarjar form. The tests access the flags directly from source and since the flags that are being accessed are not available due to the jarjar rules, there is `android.platform.test.flag.util.FlagReadException` exception.

Add jarjar rule for the tests so that tests can access the flags directly from source.

Bug: 351312723
Test: atest --test-mapping sdksandbox/TEST_MAPPING
Flag: EXEMPT Just adding jarjar rules for test
Change-Id: Ib98f8791d5a07d2f46d47241a30de3543711a534