Attempt to make test less flaky by changing the termination condition for log collection
Recent test runs of AppSearchDataMigrationHostTest have been flaky. The reason appears to be that the binder timeout is set to 5 seconds, but consent migration only starts happening close to that timeout. At 5 seconds after the getMeasurementStatus API call is made, the binder connection times out and executes the callback, which then triggers the code to stop collecting logs in the background. The successful consent migration happens *after* this, so the log messages regarding successful consent migration are absent from the collected logs, causing the test to fail.

Changing the log-collection termination condition to be independent of the measurement API call status in the hope that it'll improve the test flakiness.

Bug: 352558684
Test: CtsAdServicesHostTests
Flag: TEST_ONLY
Change-Id: I0910fc52ba1e7524d9f33ae8d81cb759a55eb450