Added methods on JobLockHolder to run with lock.
Currently, that class requires the caller to explicitly try to lock
and run something, which is error prone - this CL add methods that
take a runnable (or callable) and "wrap" it under the lock.

Besides being safer, it makes it easier to handle issues (like
logging when the lock cannot be acquired).

This CL only adds the new methods (and its tests), the existing jobs
will be refactored in future CLs (so they could be easily reverted
in case of breakages).

Test: atest AdServicesServiceCoreMeasurementUnitTests:JobLockHolderTest
Bug: 354007915
Flag: EXEMPT (bug 337358613)

Change-Id: I65ff2e0b3d94027170ba4c46ee8f1c827a2bb108