cts: always kill and recreate sandbox isolate
Make sure the JavascriptEngine isolate is killed after test execution
to ensure it is in a healthy state before the test. In the past we have
found that the isolate singleton can get killed due to memory pressure,
so avoid that problem by always using a fresh isolate instance.

Bug: 355764271
Flag: TEST_ONLY
Test: atest GenerateInputForEncodingCommandTest
Change-Id: Ic65ad1da275175067163d131dad5fdd9ced63ad9