Linter to avoid explicitly mocking ErrorLogUtil.
@SpyStatic(ErrorLogUtil.class) and @MockStatic(ErrorLogUtil.class) should
no longer be defined in test classes. ErrorLogUtil class is spied
automatically for all AdServicesExtendedMockitoTestCase subclasses.

A new linter in a different CL will be developed to ensure ErrorLogUtil
invocations are not being mocked in the test logic itself. This CL only
focuses on disallowing annotation usage.

Bug: 355699778
Test: atest ErrorLogUtilStaticMockSpyUsageDetector
Flag: TEST_ONLY
Change-Id: I39a76afdaa7b4908d109b349ecdd2ce941c6afaf