Let BackgroundFetchJob Complete Gracefully during SPE Migration
It's not necessary to be aggressive for SPE migration. Since all entry point of job scheduling is migrated to using SPE and adservice process is not a persistent process, it's okay to wait for CA service to create again then schedule the job using SPE. Moreover, we have job scheduling metrics now which can monitor the actual scheduling status.

Flag: EXEMPT (bug 337358613)
Bug: 335489810
Test: atest BackgroundFetchJobServiceTest BackgroundFetchJobTest
Change-Id: I1c921e4d00487d77bf7ad90b5ba312cf641d022f