Use Default JobPolicy in SPE When it Fails to Merge Policies
We notice some CEL SPE_INVALID_JOB_POLICY_SYNC, which complains in the method "mergeTwoJobPolicies". This method is to merge the job constraints in the binary with the constraints from the server.

This CL is to address the issue of returning empty constraints when "mergeTwoJobPolicies" throws. In this case, the job should be scheduled with the default  constraints.

Flag: EXEMPT (bug 337358613)
Bug: 359636820
Test: atest PolicyJobSchedulerTest
Change-Id: I3cf1d5a66d8f18e7177476dded5ffd08a832276e