Enable subdomain enrollment match for event reporting
Also make testFledgeFlowSuccessWithMockServer_DoesNotReportToBuyerWhenEnrollmentFails
more robust as it wasn't able to catch this issue. Previously, the test
was flaky because it was mocking the wrong interaction call and it was
only passing most of the time because it didn't properly check if the
report interaction call didn't happen before the test ends.

Test: atest com.android.adservices.service.common.FledgeE2ETest#testFledgeFlowSuccessWithMockServer_DoesNotReportToBuyerWhenEnrollmentFails
Test: atest AdServicesServiceCoreProtectedAudienceUnitTests
Flag: TEST_ONLY
Fixes: 359885280
Fixes: 357607488
Change-Id: I103e4130984ed6b57d8ec61d7930d82d0eb0cad4