Fixed ProcessLifeguardRule so test failures are always thrown.
Also attempted to fix ProcessLifeguardTestSuite - the 2nd delayed
message (to release the SyncCallback) wouldn't be sent after the
first throw (not sure why, but it's irrelevant), hence the call to
assertCalled() would fail Test2 - but prior to this change, that
issue was hidden because Test2 was failing due to bg exception
thrown by Test1 - the solution was to use non-Handler threads
instead. And the rules were set with FORWARD mode, not FAIL -
that was uber wrong. With these fixes, the suite pass locally,
but it's failing / crashing in the cloud...

Test: atest AdServicesSharedLibrariesUnitTests:ProcessLifeguardTestSuite
Fixes: 362381101
Bug: 340959631
Flag: TEST_ONLY

Change-Id: Ieb0a97c71df37901b3ce307e3dfbeb7954ad9666